# perf: use more fast and lightweight base58 conversion