@jjz 2018-09-28
Why not use base58?