@fffreedom 2018-09-28
@detailyang. 
I carefully compared the C code and your code, there are many difference between them and I am not sure it's right or not of your code in short time. 

Now, we have many other more important things(chain, transaction pool and etc.)to do and don't have enough time to verify this pr. Of course, we will verify it after October.  

Thanks a lot!